Use `ListBasedDictionary` in `ExceptionPropertiesBag`
